# gateway/lab-exchange-api-spec.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Lab Exchange API",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "https://api.prod.labdao.xyz/"
      },
      {
        "url": "https://api.stg.labdao.xyz/"
      }
    ],
    "paths": {
...
      "/datafiles": {
        "post": {
          "operationId": "addDataFile",
          "summary": "Add Data File",
          "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "walletAddress": {
                      "type": "string",
                      "format": "varchar",
                      "example": "0x1234567890abcdef1234567890abcdef12345678"
                    },
                    "filename": {
                      "type": "string",
                      "format": "varchar",
                      "example": "example_file.txt"
                    },
                    "file": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "required": [
                    "walletAddress",
                    "filename",
                    "file"
                  ]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Data file uploaded successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "cid": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad request, such as missing fields or parsing error"
            },
            "409": {
              "description": "Conflict, data file with the same CID already exists"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "get": {
          "operationId": "listDataFiles",
          "summary": "List Data Files",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "description": "Page number of the results to fetch",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 1
              }
            },
            {
              "name": "pageSize",
              "in": "query",
              "description": "Number of results per page",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 50
              }
            },
            {
              "name": "cid",
              "in": "query",
              "description": "Filter by CID",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "walletAddress",
              "in": "query",
              "description": "Filter by wallet address",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "filename",
              "in": "query",
              "description": "Filter by filename (supports partial match)",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "tsBefore",
              "in": "query",
              "description": "Filter by timestamp, to fetch records before a specific date-time",
              "required": false,
              "schema": {
                "type": "string",
                "format": "date-time"
              }
            },
            {
              "name": "tsAfter",
              "in": "query",
              "description": "Filter by timestamp, to fetch records after a specific date-time",
              "required": false,
              "schema": {
                "type": "string",
                "format": "date-time"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Paginated list of data files",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/DataFile"
                        }
                      },
                      "pagination": {
                        "type": "object",
                        "properties": {
                          "currentPage": {
                            "type": "integer"
                          },
                          "totalPages": {
                            "type": "integer"
                          },
                          "pageSize": {
                            "type": "integer"
                          },
                          "totalCount": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/datafiles/{cid}": {
        "get": {
          "operationId": "getDataFileByCid",
          "summary": "Get Data File",
          "parameters": [
            {
              "name": "cid",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Data file data",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/DataFile"
                  }
                }
              }
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Tool": {
          "type": "object",
          "properties": {
            "cid": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "walletAddress": {
              "type": "string"
            },
            "toolJson": {
              "type": "object"
            }
          }
        },
        "DataFile": {
          "type": "object",
          "properties": {
            "cid": {
              "type": "string"
            },
            "walletAddress": {
              "type": "string"
            },
            "filename": {
              "type": "string"
            },
            "jobInputs": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Job"
              }
            },
            "jobOutputs": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Job"
              }
            },
            "tags": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Tag"
              }
            },
            "timestamp": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "Flow": {
          "type": "object",
          "properties": {
            "cid": {
              "type": "string"
            },
            "jobs": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Job"
              }
            },
            "name": {
              "type": "string"
            },
            "walletAddress": {
              "type": "string"
            }
          }
        },
        "Job": {
          "type": "object",
          "properties": {
            "bacalhauJobID": {
              "type": "string"
            },
            "state": {
              "type": "string"
            },
            "error": {
              "type": "string"
            },
            "walletAddress": {
              "type": "string"
            },
            "toolID": {
              "type": "string"
            },
            "tool": {
              "$ref": "#/components/schemas/Tool"
            },
            "flowID": {
              "type": "string"
            },
            "flow": {
              "$ref": "#/components/schemas/Flow"
            },
            "inputs": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/DataFile"
              }
            },
            "outputs": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/DataFile"
              }
            }
          }
        },
        "Tag": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "dataFiles": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/DataFile"
              }
            }
          }
        }
      }
    }
  }